requirements.txt
av==11.0.0
comet-ml==3.38.1
gymnasium==0.29.1
gymnasium-robotics==1.2.4
hydra-core==1.3.2
ipykernel
matplotlib==3.8.3
mujoco==2.3.7
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
torchrl @ git+https://github.com/pytorch/rl
tensordict @ git+https://github.com/pytorch/tensordict
tqdm==4.66.2
torchinfo==1.8.0
glfw==2.7.0
pillow==10.2.0
h5py==3.10.0